requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary
python-decouple
django-jazzmin
